# src/docs/swagger.yaml
openapi: 3.0.0
info:
  title: CareerHub API
  description: The project aims to design and implement a software product using state-of-the-art tools and technologies in the software industry. The required product should mimic some functionalities of LinkedIn 
  version: 1.0.0
...
paths:
...
  /api/connections/users/{userId}:
    post:
      tags:
        - Connections
      summary: Send Connection Request
      description: Sends a connection request to another user.
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Connection request sent successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string

  # 4. Accept Connection Request
...
  /api/following/users/{userId}:
    post:
      tags:
        - Following
      summary: Follow a User
      description: Follows another user.
      parameters:
        - in: path
          name: userId
          required: true
          schema:
            type: string
      responses:
        "200":
          description: User followed successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
    delete:
      tags:
        - Following
      summary: Unfollow a User
      description: Unfollows a user.
      parameters:
        - in: path
          name: userId
          required: true
          schema:
            type: string
      responses:
        "200":
          description: User unfollowed successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
...
  /api/search/users:
    get:
      tags:
        - User Search
      summary: Search Users
      description: Searches for users based on name, company, or industry.
      parameters:
        - in: query
          name: query
          required: true
          schema:
            type: string
          description: Search term for user name, company, or industry.
        - in: query
          name: company
          required: false
          schema:
            type: string
          description: Filter by company name.
        - in: query
          name: industry
          required: false
          schema:
            type: string
          description: Filter by industry.
        - in: query
          name: page
          required: false
          schema:
            type: integer
          description: Page number for pagination.
        - in: query
          name: limit
          required: false
          schema:
            type: integer
          description: Number of results per page.
      responses:
        "200":
          description: Users retrieved successfully
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    id:
                      type: string
                    firstName:
                      type: string
                    lastName:
                      type: string
                    headline:
                      type: string
...
  /api/reports/users/{userId}:
    post:
      tags:
        - Privacy Controls
      summary: Report a User
      description: Reports a user for inappropriate behavior or content.
      parameters:
        - in: path
          name: userId
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                reason:
                  type: string
                details:
                  type: string
      responses:
        "200":
          description: User reported successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
...